+++ CHANGELOG.md
@@ -8,7 +8,7 @@ log is based on the [Keep a CHANGELOG](http://keepachangelog.com/) project.
 
 ## Updated
 
-- Include predictive failure label to physical disk metrics [#137](https://github.com/Comcast/fishymetrics/issues/137)
+- Include predictive failure label to SATA and NVMe disk metrics [#137](https://github.com/Comcast/fishymetrics/issues/137)
 
 ## [0.15.0]